The application now minimises to the system tray.
Fullscreen now exits when the video has finished playing.

If you do not have the "Start videos Fullscreen" option selected, the application will restore itself from the system tray when a video is played. If you do have "start videos fullscreen" checked the app will remain in the system tray, and only the video will fullscreen - this means you can leave the app constantly in the system tray, and it will fullscreen when you start a video, and exit fullscreen when it stops, without the app ever being visible - making the application less apparent and less interfering. Videos will just fullscreen and exit fullscreen as needed.

If you have the application minimised and you stream a picture, the application will restore itself to show the picture. 

Added an icon that's used for the system tray and app icon TODO: make the background of the icon transparent

git-svn-id: http://airstream-media-player.googlecode.com/svn/trunk@25 0d03175f-7565-ef8f-9c92-1ae2cd87759a